# openapi.yaml
openapi: "3.2.0"
info:
  title: sing-box hub
  version: 1.3.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    独立的多节点 sing-box 监控管理面板。

    - **数据面**:复用每个 sing-box 节点内置的 Clash API(`experimental.clash_api`)采集
      实时流量、连接(含客户端来源 IP)与内存信息;节点侧无需安装任何组件。
    - **能力**:多节点注册与健康检查、流量历史(分钟/小时/天)、按来源 IP 的流量统计、
      出口 IP 探测、基本控制(断开连接、切换节点选择、切换 mode)、订阅导出
      (sing-box JSON / Clash YAML / base64 分享链接 URI 列表)。
    - **约定**:错误一律 RFC 9457 Problem Details(`application/problem+json`);
      列表游标分页(`cursor` + `page_size`,响应 `next_cursor` 与绝对 URI `links`);
      时间 RFC 3339 UTC;字段 snake_case;数量字段带单位后缀(`_bytes`、`_bps`、`_ms`)。
    - **订阅安全取舍**:`GET /sub/{token}` 为「能力 URL」(capability URL),令牌置于路径而非
      Authorization 头,原因是订阅客户端 App 无法携带自定义请求头。缓解:令牌可随时轮换与吊销、
      端点只读、有限流与访问日志、建议部署侧强制 HTTPS。
    - **节点不可达**:面板到节点的 Clash API 请求失败统一返回 `502`,
      `type: https://sing-hub.dev/probs/node-unreachable`。
    - **加密配置导入**:节点可选 `config_url` + `kce_key`;添加/更新时面板自动拉取
      (支持 KCE1 加密封装)并解析 Clash 代理,全部代理自动导入为面板节点;上游
      拉取/解密失败统一返回 `502`,`type: https://sing-hub.dev/probs/upstream-unavailable`。
    - **账号与会话**:所有面板账号均有管理权限。访问令牌绑定数据库中的会话;
      删除账号、修改密码或用户名、登出会撤销相关会话。刷新令牌只能消费一次。
      升级前的访问令牌需要刷新,有效的旧刷新令牌仍可换取新会话。
    - **公开状态**:`/api/public/*` 无需登录,公开节点名称、标签、出口 IP、版本和运行统计,
      不返回 API 地址、API secret、来源 IP 明细或出站配置。
servers:
  - url: http://localhost:9090
    description: 本地开发默认地址
...
security:
  - bearerAuth: []
paths:
...
  /nodes/{node_id}:
    parameters:
      - { $ref: "#/components/parameters/nodeId" }
    get:
      operationId: getNode
      summary: 节点详情
      tags: [nodes]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateNode
      summary: 部分更新节点(application/merge-patch+json)
      tags: [nodes]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/NodePatch" }
            examples:
              rename:
                value: { name: vps-jp-2, remark: moved }
              rotateSecret:
                value: { api_secret: new-s3cret }
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: deleteNode
      summary: 删除节点(连同其统计数据)
      tags: [nodes]
      responses:
        "204": { description: 已删除 }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录 /session 获取的访问令牌
  parameters:
...
    nodeId:
      name: node_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationError:
      description: 语义校验失败
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    PreconditionFailed:
      description: If-Match 不匹配
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details;type 为稳定绝对 URI,未知类型为 about:blank
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errors:
          type: array
          description: 校验错误明细,pointer 为 RFC 6901 JSON Pointer
          items:
            type: object
            required: [detail, pointer]
            properties:
              detail: { type: string }
              pointer: { type: string }
      examples:
        - type: https://sing-hub.dev/probs/node-unreachable
          title: Node unreachable
          status: 502
          detail: "dial tcp 192.168.1.10:9090: connect refused"
          instance: /nodes/3fa85f64-5717-4562-b3fc-2c963f66afa6/connections
...
    ImportStats:
      type: object
      required: [found, imported, updated, removed, skipped]
      properties:
        found: { type: integer, description: 本次解析到的代理数 }
        imported: { type: integer, description: 新导入节点数 }
        updated: { type: integer, description: 刷新出站的既有节点数 }
        removed: { type: integer, description: 上游已消失而删除的节点数 }
        skipped: { type: integer, description: 因类型不支持或字段缺失被跳过的代理数 }
    Node:
      type: object
      required: [id, name, api_url, api_secret_set, tags, enabled, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        api_url: { type: string, format: uri }
        api_secret_set:
          type: boolean
          description: 已配置密钥则为 true;密钥本身永不回读
        outbound_json:
          type: [string, "null"]
          description: 节点出站配置(sing-box outbound JSON);列表接口不返回此字段
        source:
          type: string
          description: 节点来源;''=手工添加,`sub:<id>`=由上游订阅源导入
        config_url:
          type: string
          description: 加密配置地址(KCE1 密文或明文 Clash YAML);为空表示未配置解密导入
        kce_key_set:
          type: boolean
          description: 已保存 KCE1 解密密钥则为 true;密钥本身永不回读
        import:
          description: 仅创建/更新且携带 config_url 时返回的解密导入统计
          allOf:
            - $ref: "#/components/schemas/ImportStats"
        tags:
          type: array
          items: { type: string }
        enabled: { type: boolean }
        remark: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_online_at: { type: [string, "null"], format: date-time }
...
    NodePatch:
      type: object
      description: merge-patch 语义;未提供的字段保持不变
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        api_url: { type: string, format: uri }
        api_secret: { type: string, minLength: 1 }
        outbound_json: { type: [string, "null"] }
        config_url: { type: [string, "null"] }
        kce_key: { type: string }
        tags:
          type: array
          items: { type: string }
        enabled: { type: boolean }
        remark: { type: string }
...